First draft of FusedConvMaxpoolOp.
Currently supports ksize but assumes the stride is on the window
boundaries. This op allows us to max pool directly after convolution
without maintaining a temporary intermediate tensor.

```
for each maxpool location:
    block = expand location to input space
    max_val = 0
    for each (px, py) in block:
        max_val = max(max_val, local_conv_filter(px,py,...)
```

TODO: The convolution operator computes the filter tap offsets at the
start of the runthrough using the output sizes. I need to make sure that
this isnt broken.